requirements.txt
# ============================================================================
# FILE: requirements.txt (Python Dependencies)
# ============================================================================

"""
flask==2.3.3
flask-cors==4.0.0
pywin32==306
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
"""
